requirements.txt
# Modern Data Stack - Airflow Native
apache-airflow==2.10.4
astronomer-cosmos[dbt-postgres]==1.8.0
dlt[postgres]==1.7.0
dbt-core==1.8.8
dbt-postgres==1.8.2

# Data Processing
pandas==2.1.4
numpy>=1.21.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.35

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.10.4

# Additional dependencies
pyarrow>=10.0.0
requests>=2.28.0
kagglehub>=0.2.0